Refactor : Move method/field
